README.md
`SCHOEN Peek` is the first node I created for an industrial design tool-set. It´s a versatile input node for ComfyUI that allows you to capture any area of your screen in real-time and use it as an image input for your workflows. 
It is ideal for interactive image generation, live feedback loops, or directly integrating external content like videos, designs, or games into your AI workflows.